feat(lattice-list): add local id and global id concept to lattice list
we add local id and global id concept to lattice list. local id of a lattice is based on ghost
boundary of lattice list and global id is based on global simulation box boundary. We can calculate
global id of all lattice in simulation box of current process, and local lattice id of all lattice
in ghost area and simulation area of current process.