fix(rate): pass id of source and target lattice to RatesSolver::deltaE to replace Lattice::getId
we pass source lattice id and target lattice id to RatesSolver::deltaE member function. In preview
implementation, the source lattice id and target lattice id are got by calling Lattice::getId method
on Lattice object, which is correct for lattices in simulation area but wrong for lattices in ghost
area to get lattice coordinate in lattice-array.