fix(model, kmc): correct preview bug fix `dont insert defect to defect-map if the new defect is in ghost area`
In preview bug fix, it use old_lat_id instead of new_lat_id to check whetherm the
target lattice in selcted event(new defeect) is in ghost area, which is not correct.
In this change, we fix this mistake by using new_lat_id as parameter of function
`LatListMeta::isGhostLat`.